Store BBLAYERS and local.conf fragments in specification order again
According to c9326cc1edc9, we had issues with producing a deterministic
order for those config elements. As sorting is a rather blunt method of
resolving it and as ordering plays a role in bitbake's processing of the
configs, we should try harder to provide more user control over that.

A key role in this play ordered dictionaries that report their content
the same order it was added. We already use OrderedDict for merging the
config dict across includes, see _internal_dict_merge. Includes are
merged recursively, depths-first, thus are expected to be deterministic
and predictable. We just need to convert building the repo_dict to
OrderedDict, and them sorted() can be removed.

Closes: https://github.com/siemens/kas/issues/36

Reported-by: Marius Kriegerowski <marius.kriegerowski@gfz-potsdam.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>